requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
